% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seamsCFR-package.R
\docType{package}
\name{seamsCFR-package}
\alias{seamsCFR}
\alias{seamsCFR-package}
\title{seamsCFR: SEAMS Demo: Lagged CFR computation}
\description{
Provides a demo solution to the reuse practical.
}
\author{
\strong{Maintainer}: Carl Pearson \email{carl.ab.pearson@gmail.com} (\href{https://orcid.org/0000-0003-0701-7860}{ORCID})

}
\keyword{internal}
